feat: refactor schema management by splitting core functionalities into dedicated modules
This commit introduces a significant refactor of the schema management system by splitting the core functionalities from `core.rs` into several dedicated modules, enhancing maintainability and organization. Key changes include:

- Created new modules: `schema_operations`, `schema_state_management`, `schema_persistence`, and `schema_validation` to encapsulate specific functionalities.
- Moved relevant methods from `SchemaCore` in `core.rs` to the newly created modules, ensuring a clearer separation of concerns.
- Updated `mod.rs` to re-export the new modules, maintaining accessibility for existing functionality.
- Removed redundant code and improved documentation for clarity.

These changes aim to improve the structure of the schema management system, making it easier to navigate and extend in the future.
