+++ src/schema/schema_validation.rs
@@ -1,0 +1,2 @@
+// This module contains validation-related methods that are implemented elsewhere
+// The actual implementations are in the transform.rs and persistence.rs modules
